+++ csv_generator/Cargo.toml
@@ -9,3 +9,10 @@ edition = "2021"
 random_name_generator = "0.3.6"
 csv = "1.3.0"
 rand = "0.8.5"
+
+[dev-dependencies]
+criterion = "0.5.1"
+
+[[bench]]
+name = "csv_generator_benchmark"
+harness = false
